feat: added `Message Broker` integration in Drifty GUI classes removing direct method calls to `setLinkOutput`, `setDirOutput`, `setFilenameOutput` and `setDownladOutput` methods.
I have also added a logger instance to log the start of the GUI program

+++ src/main/java/GUI/Forms/GetFilename.java
@@ -112,9 +112,10 @@ public class GetFilename extends Task<ConcurrentLinkedDeque<Job>> {
                 }
                 for (File file : deleteList) {
                     try {
-                        FileUtils.forceDelete(file);
-                    } catch (IOException ignored) {
-                    }
+                        if (file.exists()) {
+                            FileUtils.forceDelete(file);
+                        }
+                    } catch (IOException ignored) {}
                 }
             }
         };
